avera: fix and rename to avera_us
666 locations are returned by a new API endpoint, but 100 of these are
missing coordinates.
